'advanced_voice_model' and 'voice_model' usage for xTTS need xTTS API Mantella server 1.3 (#236)
* tts.py change voice update

change to change voice so it doesn't try to change model at each speaker change, the feature is still there

* Fixed typo

* 'advanced_voice_model' and 'voice_model' usage for xTTS

The character_df is passed to the tts.Synthesizer to build two dictionaries for xTTS based on 'advanced_voice_model' and 'voice_model', which are compared to the return value of xtts_get_speakers_list.

I implemented this so that when there is a need to search for 'advanced_voice_model' or 'voice_model', it will not browse the entire CSV values but only those that can be present inside the xTTS /speakers folder and /latent_speaker_folder.

Additionally, two .csv files are created in the root folder that lists the used/not used speakers across 'advanced_voice_model' and 'voice_model' based on the comparison that occurs at the start.

It needs to use the new xTTS API Mantella server 1.3 to work.

I've cleaned up the way it changes voice and removed some unnecessary calls.

I did test xVASynth this time, so it should be okay!

* update _get_available_models

_get_available_models to lowercase

* Changed skyrim_voice_folder to more generic game_voice_folder

* Removed duplicate code

* Added empty advanced_voice_model input for generic NPCs

* Moved XTTS logs to data folder