Add sentence accumulator (#545)
* Add sentence accumulator

Adds a sentence accumulator that adds a guard in front of the parsing_chain.
LLM output that is too short to be processed correctly will not be handed to the chain until something is accumulated that can be parsed.
LLM output that is too long gets pre cut into correct chunks before handed to the chain.

* Moved clean sentence into the sentence_accumulator

* Add test for Gemini-like token outputs

---------